move essence parser to its own crate
