+++ conjure_oxide/Cargo.toml
@@ -11,6 +11,7 @@ walkdir = "2.5.0"
 [dependencies]
 conjure_core = { path = "../crates/conjure_core" }
 minion_rs = { path = "../solvers/minion" }
+conjure_essence_parser = { path = "../crates/conjure_essence_parser" }
 
 
 uniplate = "0.2.2"
@@ -43,7 +44,7 @@ tracing-subscriber = { version = "0.3.18", features = [
 tracing = "0.1.41"
 tree-sitter = "0.24.7"
 tree-sitter-essence = { version = "0.1.0", path = "../crates/tree-sitter-essence" }
-tree-morph = { path = "../crates/tree_morph"}
+tree-morph = { path = "../crates/tree_morph" }
 tree-sitter-haskell = "0.23.0"
 git-version = "0.3.9"
 project-root = "0.2.2"
