+++ conjure_oxide/src/utils/conjure.rs
@@ -13,80 +13,14 @@ use serde_json::{Map, Value as JsonValue};
 use itertools::Itertools as _;
 use tempfile::tempdir;
 
-use thiserror::Error as ThisError;
-
-use crate::model_from_json;
+use crate::parse_essence_file;
 use crate::solver::adaptors::Minion;
 use crate::solver::Solver;
 use crate::utils::json::sort_json_object;
-use crate::Error as ParseErr;
+use crate::EssenceParseError;
 use crate::Model;
 
 use glob::glob;
-
-#[derive(Debug, ThisError)]
-pub enum EssenceParseError {
-    #[error("Error running conjure pretty: {0}")]
-    ConjurePrettyError(String),
-    #[error("Error running conjure solve: {0}")]
-    ConjureSolveError(String),
-    #[error("Error parsing essence file: {0}")]
-    ParseError(ParseErr),
-    #[error("Error parsing Conjure solutions file: {0}")]
-    ConjureSolutionsError(String),
-    #[error("No solutions file for {0}")]
-    ConjureNoSolutionsFile(String),
-}
-
-impl From<ParseErr> for EssenceParseError {
-    fn from(e: ParseErr) -> Self {
-        EssenceParseError::ParseError(e)
-    }
-}
-
-pub fn parse_essence_file(
-    path: &str,
-    filename: &str,
-    extension: &str,
-    context: Arc<RwLock<Context<'static>>>,
-) -> Result<Model, EssenceParseError> {
-    parse_essence_file_1(&format!("{path}/{filename}.{extension}"), context)
-}
-
-fn parse_essence_file_1(
-    path: &str,
-    context: Arc<RwLock<Context<'static>>>,
-) -> Result<Model, EssenceParseError> {
-    let mut cmd = std::process::Command::new("conjure");
-    let output = match cmd
-        .arg("pretty")
-        .arg("--output-format=astjson")
-        .arg(path)
-        .output()
-    {
-        Ok(output) => output,
-        Err(e) => return Err(EssenceParseError::ConjurePrettyError(e.to_string())),
-    };
-
-    if !output.status.success() {
-        let stderr_string = String::from_utf8(output.stderr)
-            .unwrap_or("stderr is not a valid UTF-8 string".to_string());
-        return Err(EssenceParseError::ConjurePrettyError(stderr_string));
-    }
-
-    let astjson = match String::from_utf8(output.stdout) {
-        Ok(astjson) => astjson,
-        Err(e) => {
-            return Err(EssenceParseError::ConjurePrettyError(format!(
-                "Error parsing output from conjure: {:#?}",
-                e
-            )))
-        }
-    };
-
-    let parsed_model = model_from_json(&astjson, context)?;
-    Ok(parsed_model)
-}
 
 pub fn get_minion_solutions(
     model: Model,
@@ -252,7 +186,7 @@ pub fn get_solutions_from_conjure(
     let mut solutions_set = vec![];
     for solutions_file in solutions_files {
         let solutions_file = solutions_file.unwrap();
-        let model = parse_essence_file_1(solutions_file.to_str().unwrap(), Arc::clone(&context))
+        let model = parse_essence_file(solutions_file.to_str().unwrap(), Arc::clone(&context))
             .expect("conjure solutions files to be parsable");
 
         let mut solutions = BTreeMap::new();
