+++ conjure_oxide/src/utils/mod.rs
@@ -1,5 +1,4 @@
 pub mod conjure;
-pub mod essence_parser;
 pub mod json;
 pub mod misc;
 pub mod testing;
