+++ crates/conjure_essence_parser/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "conjure_essence_parser"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+conjure_core = { path = "../conjure_core" }
+tree-sitter-essence = { path = "../tree-sitter-essence" }
+uniplate = "0.2.2"
+tree-sitter = "0.24.7"
+thiserror = "2.0.12"
+
+
+[lints]
+workspace = true
