+++ crates/conjure_essence_parser/src/lib.rs
@@ -1,0 +1,7 @@
+mod errors;
+mod parser;
+mod parser_legacy;
+
+pub use errors::EssenceParseError;
+pub use parser::*;
+pub use parser_legacy::parse_essence_file;
