+++ crates/conjure_essence_parser/src/parser/letting.rs
@@ -1,0 +1,51 @@
+#![allow(clippy::legacy_numeric_constants)]
+use std::collections::BTreeSet;
+use std::rc::Rc;
+
+use tree_sitter::Node;
+
+use conjure_core::ast::Declaration;
+use conjure_core::ast::{Name, SymbolTable};
+
+use super::domain::parse_domain;
+use super::expression::parse_expression;
+use super::util::named_children;
+
+/// Parse a letting statement into a SymbolTable containing the declared symbols
+pub fn parse_letting_statement(letting_statement_list: Node, source_code: &str) -> SymbolTable {
+    let mut symbol_table = SymbolTable::new();
+
+    for letting_statement in named_children(&letting_statement_list) {
+        let mut temp_symbols = BTreeSet::new();
+
+        let variable_list = letting_statement.child(0).expect("No variable list found");
+        for variable in named_children(&variable_list) {
+            let variable_name = &source_code[variable.start_byte()..variable.end_byte()];
+            temp_symbols.insert(variable_name);
+        }
+
+        let expr_or_domain = letting_statement
+            .named_child(1)
+            .expect("No domain or expression found for letting statement");
+        match expr_or_domain.kind() {
+            "expression" => {
+                for name in temp_symbols {
+                    symbol_table.insert(Rc::new(Declaration::new_value_letting(
+                        Name::UserName(String::from(name)),
+                        parse_expression(expr_or_domain, source_code, &letting_statement_list),
+                    )));
+                }
+            }
+            "domain" => {
+                for name in temp_symbols {
+                    symbol_table.insert(Rc::new(Declaration::new_domain_letting(
+                        Name::UserName(String::from(name)),
+                        parse_domain(expr_or_domain, source_code),
+                    )));
+                }
+            }
+            _ => panic!("Unrecognized node in letting statement"),
+        }
+    }
+    symbol_table
+}
